+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [process][linux] implement some oneshot functions
+
 ## [v3.0.1] - 2020-02-12
 
 ### Fix